Improved SceneLoader, Player & HUD
Improve the SceneLoader so now it has a list with all active scenes. So it can unload them later. Now instead of hard coding the scenes index or name you can serialize the scenes Index & Name thanks to the class SceneRef that allows you to pass a scene asset in editor and keep the name & index.

Also now the player script has new events which allows the HUD (for example) to know when should he be updated instead of updating it self every frame